+++ crates/shielded-pool/src/component/note_manager.rs
@@ -59,7 +59,7 @@ pub trait NoteManager: StateWrite {
 
         let note = Note::from_parts(*address, value, Rseed(rseed_bytes))?;
         // Now record the note and update the total supply:
-        self.update_token_supply(&value.asset_id, value.amount.value() as i64)
+        self.update_token_supply(&value.asset_id, value.amount.value() as i128)
             .await?;
         self.add_state_payload(StatePayload::Note {
             note: note.payload(),
